RFID: Update definitions to allow a second reader
The board has a second SPI bus wired up for communication with two RFID readers
if needed.
However, the code only provided a constructor with no arguments, preventing
the use of a second instance on the different bus.

Update the constructor and the definitions to allow changing the chip select
and SPI bus used when creating an RFID reader.